functions.c update: variable was changed

+++ functions.c
@@ -12,7 +12,7 @@ int convert_base(int base, int number, int band)
 	char buffer[17];
 	char buffersito[20];
 	char *ptr = &buffersito[20];
-	int save = number;
+	long int save = number;
 
 	if (band == 0)
 		_strcpy(buffer, "0123456789ABCDEF");
